Disable provisioning API after privisioning process complete
The handle_net_manager_request() function checks if the provisioning API
is enabled and if the device is provisioned before handling any requests,
but the api_enabled flag wasn't being updated appropriately. This fix
triggers a re-evaluation of the api_enabled flag.

Additionally, the handle_net_manager_request() function was calling the
send_prov_response() function and passing the cur_net_req_obj (type of
'dict') when it expects the first parameter to be of type 'int'. This has
been updated to call the send_response() function instead.

Bug 19599