Implement reporting of network connection stats via DBus
As part of the efforts to support device management (EdgeIQ), the IG60
needs to be able support reporting (and eventually remote configuring)
of network interface settings, including enterprise Wi-Fi (802.1x)
settings.

This change adds a new DBus signal to the Config service which is
triggered by the NetworkManager properties changing. When this occurs,
the basic stats of the Ethernet and Wi-Fi interfaces are collected and
exposed via the new ConnectionStatsChanged signal.

The EdgeIQ client supports a 'DBus Signal Listener' 'Ingestor' which
can be configured to listen for the new ConnectionStatsChanged signal
and to subsequently relay the new stats up to the EdgeIQ cloud.

Bug 19966