netstat: schedule the status updates so only the last one is sent
EdgeIQ has a feature to listen for a configurable D-Bus signal, and we use this
feature in conjunction with our Config service. The Config service listens for
property changes from NetworkManager, filters and repackages the data, and emits
it's own D-Bus signal which is picked up by the EdgeIQ client.

NetworkManager can be 'verbose' and emits many signals when an interface is going
up or down or simply changing. This causes a backlog of messages through our data
pipeline to EdgeIQ and can result in the messages reaching the EdgeIQ dashboard
out of order.

This change introduces a delay before sending the status message. If another
property change signal is received before the delay, the original message is
canceled, and the timer is restarted. This ensures that only the last status
message is sent.

Bug 20405